feirinha-api-clean/requirements.txt
blinker==1.8.2
click==8.1.8
Flask==3.0.3
flask-cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.0.6
gunicorn==21.2.0
gspread==6.0.2
google-auth==2.28.1
